fix(pipelines): self-mutate always adds analytics (#19010)
When we moved analytics generation from the CLI to the framework, we
kept a fallback in the CLI for backwards compatibility: if there were no
analytics in the templates but analytics were requested, they would
still be added in the CLI.

This breaks in the case of CDK Pipelines self-mutate step, if analytics
have been disabled: there are no analytics in the template, and the
default setting for analytics is `true`, since there is no `cdk.json`
to opt out, so they will always be added back in.

A better solution to check whether the CLI needs to do the fallback
is to go off of the cx schema version number: the CLI can know that
it never needs to do anything for cloud assemblies generated starting
at a certain version.

Fixes #18933.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*