docs(elasticloadbalancingv2): clarify the behavior of the `open` listener property (#19019)
The documentation for the `open` property leads to some confusion on the
actual behavior. Setting `open` to `true` will not open up the listener
since listeners do not have security groups. It will instead open up the
load balancer on the listener port.

There is a bug report #6281 that says this property should be moved to
the load balancer, but I think the current behavior is the intended
behavior and that the current documentation just makes it confusing.
The intention of this property is to give an easy way to add a security
group rule to the load balancer security group that allows anyone to
access the load balancer on the port of the listener. If we were to move
this property to the load balancer it would become more cumbersome
because we would have to also provide the port that we want to allow
traffic on (load balancers do not have a port).

closes #6281


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*