chore: version check rejects non-monorepo package.jsons (#19033)
Because of the following confluence of historical decisions:

- Our `package.json`s are the source of truth for all our packages
- We want to minimize merge conflicts and we want to be able to
  tell when we are doing a development build.

We have a mechanism called `align-versions` that is going to update
the version `0.0.0` in all package jsons to `1.144.0` (or whatever
the current version number is) just prior to doing the build.

It will look in the Lerna/Yarn workspace config to find all
package.jsons to replace.

Afterwards, to make sure, a grep on *ALL* package.jsons makes
sure there are no stray `0.0.0`s left anywhere in the repo.

We recently added another `package.json` to help Dependabot
do version checks on packages we use in our repository: that new
`package.json` was NOT aligned (because it's not a monorepo
package itself), so it would stay at `0.0.0` and the integrity
check would fail the build.

Update the integrity check to stop considering packages that
have been marked as `private: true`. Those won't be published to
NPM anyway, and so it also doesn't matter what their version number is.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*