fix(route53): fix cross account delegation deployment dependency
For each zone to delegate a policy is created and attached to the
handler role. This change adds an explicit dependency between the policy
attachment and the lambda handler to make sure the cross account
delegation handler is not started before the policy is created and
attached to the handler role.

Signed-off-by: Philipp Hoefflin <philipp.hoefflin@haufe-akademie.de>