feat(cloudfront-origins): extend max keepaliveTimeout of HttpOrigin to 180 (#18837)
This pull request extends the maximum value of the keepaliveTimeout property for HttpOriginProps up to 180 seconds. This allows the use of the construct also in the case that a limit increase has been approved for CloudFront [origin keep-alive timeout quota](https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/distribution-web-values-specify.html#DownloadDistValuesOriginKeepaliveTimeout).

The 180 seconds cap should be AWS hard limit for that quota.

This is related to #18697, which only extended the readTimeout, but keepaliveTimeout can also be increased.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*