fix(route53): fix cross account delegation deployment dependency (#19047)
For each zone to delegate a policy is created and attached to the
handler role. This change adds an explicit dependency between the policy
attachment and the lambda handler to make sure the cross account
delegation handler is not started before the policy is created and
attached to the handler role.

fixes: #19041 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*