fix(elasticloadbalancingv2): validate port/protocol are not provided for lambda targets (#19043)
When creating a target group with the targetType = `LAMBDA` you should
not provide the port or protocol. If protocol is provided then
CloudFormation will throw an error message. If you provide the port to
CDK, CDK will figure out and provide the protocol as well.

This PR adds validation and throws an error if either port or protocol
is provided when the target type is Lambda.

fixes #12514


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*