feat(lambda): add a fromFunctionName() method (#19076)
After #18255 has been merged, the account and region of the Function's imported ARN is correctly recognized.
Unfortunately, this has surfaced a few cases where that causes problems,
like when using an imported Function as the target of a CodePipeline invoke Action
(the pipeline construct needs to verify that the target Function is in the same account and region as the pipeline).

Add a `Function.fromFunctionName()` method that allows you to import a Function in the same account and region.

Fixes #19031

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*