fix(cli): hotswapping is slow for many resources deployed at once
In the `LazyListStackResources` class,
we were incorrectly caching the results of the ListStackResources CloudFormation API call.
This resulted in executing redundant calls,
and very visibly impacted the performance of hotswapping when the Stack included many resources.

In manual testing, this change brought down deployment time of a large Stack from 80 to 7 seconds.

Closes #19021