fix: API compatibility check fails in CI pipeline
In the master build, we do a version bump which turns `1.146.0` into
`1.146.0-rc.0`, which makes the API compatibility checker try to install
packages with those versions.

Because none of those packages exist, the list of packages ends up
empty, which turns our `npm install $PACKAGES` command into `npm
install`, which subsequently complains that there's no `package.json`.

In order to make this work properly, we need to query the ACTUAL
released version numbers on GitHub so we can properly find the
version to diff against.