feat(cli): support for matching notices with arbitrary module names (#19088)
This change allows us to publish notices to match any arbitrary module or
construct. For example, a notice that targets the component

```json
{
    "name": "@aws-cdk/aws-apigatewayv2-alpha",
    "version": ">=2.10.0"
}
```

will be displayed to the user if they have in their cloud assembly a resource of
type `@aws-cdk/aws-apigatewayv2-alpha.HttpStage` or
`@aws-cdk/aws-apigatewayv2-alpha.HttpApi` or any other construct from the
`apigatewayv2` module, as long as it was built using some version of the
module above or equal to 2.10.0.

We can also target a specific construct, instead:

```json
{
    "name": "@aws-cdk/aws-apigatewayv2-alpha.HttpStage",
    "version": ">=2.10.0"
}
```


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*