fix(s3-notifications): notifications allowed with imported kms keys (#18989)
----
fixes: #18988

If you add an sqs queue as notification target to an s3 bucket, and this sqs queue is encrypted with an imported kms IKey, the stack won't synthesize. Instead of failing, it should warn the user, that it can not ensure the correct kms key policy permissions.

This fix will solve this.

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*