build: update verify-imports-resolve-same script to ignore module extensions
verify-imports-resolve-same.ts script validates if the package import for
monorepo package imports the same modules, but when validating the imports
typescript either would provide .d.ts or .ts or .js file based on how it is
imported. Since .ts and .d.ts are typescript artifacts and .js is used at
runtime the extension can be safely ignored. The code to ignore the extension
was not escaping the `.` char and string replacement was failing because of it.
Update the regex to escape `.`