fix: apply tags to nested stack
When a nested stack gets a custom tags the tag gets applied only to the resources defined in the stack and not the stack itself. This change fixes this issue and applies the tag to the nested stack as well
This was caused by NestedStack subclassing Stack but not using the NestedStack.tags in the CfnStack resource that was created by it and not rendering the tags that were added when creating the AWS::CloudFormation::Stack. This change updates the `_prepareTemplateAsset` to get all the tags added to the nested stacks instance TagManager and add them to the CfnStack before rendering the assets

fixes #17463