fix(cli): hotswapping is slow for many resources deployed at once (#19081)
In the `LazyListStackResources` class,
we were incorrectly caching the results of the ListStackResources CloudFormation API call.
This resulted in executing redundant calls,
and very visibly impacted the performance of hotswapping when the Stack included many resources.

In manual testing, this change brought down deployment time of a large Stack from 80 to 7 seconds.

Closes #19021

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*