fix: apply tags to nested stack (#19128)
When a nested stack gets a custom tags the tag gets applied only to the resources defined in the stack and not the stack itself. This change fixes this issue and applies the tag to the nested stack as well. This was caused by NestedStack subclassing Stack but not using the NestedStack.tags in the `CfnStack` (as it  accepts only `CfnTag[]` and not `TagManager`)

This change updates the `_prepareTemplateAsset` to get all the tags added to the `CfnStack`  from `NestedStack.tag`  before rendering the stack
    
fixes #17463


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*