docs(aws-events-targets): fix StepFunctions example
The role needs to be passed to the SfnStateMachine target, not to the StateMachine resource.

The role's trust policy trusts events.amazonaws.com so the state machine resource would be unable to use this role anyways. Leave the StateMachine construct to create its own role. Instead, pass the role to SfnStateMachine where EventBridge will use it to start the state machine.