fix(eks): Helm charts fail to install when provided as an asset (#19180)
When the helm chart is specified by an asset instead of a respository URL, the `repository` variable in the custom resource is `None`. Adding a check before using the variable.

This bug wasn't caught by integration tests because the existing integration test that exercises EKS clusters creates a cluster with many features, including two different helm charts: one from a URL and another one from an asset. This change includes a new integration test that exclusively deploys a cluster with a helm chart from an asset.

Fixes https://github.com/aws/aws-cdk/issues/19179.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*