fix(cli): deprecated stack ids printed at the end of synth (#19216)
Using the regular stack identifier in cli commands is deprecated in v1, and removed in v2. If you supply the `.id` of a stack to any cdk commands (`synth`, `deploy`), then v1 will print out a deprecation warning. And, v2 will not work at all. Because of this, it is confusing to users that `cdk synth` prints out the regular stack identifier. It is best to print out the hierarchical identifier in all cases. 

At minimum, this partially fixes #18599. This will fix the scenario where the CLI prints out 
```
Supply a stack id (StageOneF5E336C2, StageTwo81B557F6) to display its template.
```
when those ids will not actually work, and the hierarchical ids, `Stage/One` and `Stage/Two`, should be printed instead.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*