fix(cli): `watch` logs always end with the 'truncated' message
CloudWatchLogs.filterLogEvents will always return a `nextToken` as long
as there are _any_ additional logs. This is regardless of any filter
used (i.e. `startTime`).

This PR updates the logic to only display the truncated message if
`CloudWatchLogs.filterLogEvents` returns 100 events (the `limit`) _and_ the
`nextToken`. If the limit is hit and there is a `nextToken` then we can
assume that the _current_ request for log events was truncated.

fixes #18805