chore(aws-events-targets): fix StepFunctions doc & tests role usage (#19178)
The role used in the aws-events-targets example and test code needs to be passed to the SfnStateMachine target, not to the StateMachine resource.

The role's trust policy trusts events.amazonaws.com so the state machine resource would be unable to use this role anyways. This PR modifies the example in README.md and the test code to have the StateMachine construct create its own role and pass the manually created role to SfnStateMachine where EventBridge will use it to start the state machine.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*