feat(elasticsearch): Decouple setting access policies from domain constructor (#15876)
Currently when creating an elasticsearch domain the access policies must be set in the constructor. This makes it impossible to create access policies that reference the domain. 

### Use Cases

See: https://aws.amazon.com/premiumsupport/knowledge-center/kinesis-firehose-cross-account-streaming/

### Proposed Solution

This PR extracts the access policy setting to a helper method which can be used if the `accessPolicies` and `useUnsignedBasicAuth` props are not set. The helper will error if access policies are already set to prevent creating duplicate custom resources.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*