chore: make node-bundle's "build" script behave as others (#19307)
The default `projen` build step re-generates the `package.json` file and
re-install dependencies, which breaks in case the `align-versions.sh`
script was run (as the re-generated `package.json` has version `0.0.0`
again).

This inserts a hack to reset the `build` script and make it a synonym
to the `compile` script.

This was breaking the `jsii` integration tests (they try to build, but
not test, `aws-cdk`).