chore(cli): reorganize CLI code
Reorganize CLI code so that the `aws-auth` and `plugin` directories can
be (mostly) airlifted out. Any dependencies it has on surrounding code
is encoded in the `_env.ts` source file, which will do foreign imports.

The plugin subsystem used to call the context provider subsystem
directly--that's now inverted so that you don't have to lift the context
provider system out with it.