chore(release-check): add retries (#19336)
The `find-latest-release` script occasionally times out on PR builds.

Add sleeps and retries. The retries are pretty hefty because the API
limits are quite low. We can't just toss a GitHub Token in there, as
anyone who submits a PR would have access to it. This is the cheapest
fix for now.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*