chore: allow opt-out of test.concurrent in integ tests
The `test.concurrent` feature of `jest` immediately executes the test
code in order to obtain a `Promise` that is checked as part of the test
run.

This means that the test handler code runs whether the test is included
or not in the assessment, as controlled by the `-t` flag of jest.

In order to allow integration tests to selectively run only a subset of
the tests, we must hence not use `test.concurrent`, or else all test
handlers run fully, leading to unwanted side effects.

This is needed in order to be able to complete cdklabs/cdk-ops#1922.