docs: add integration testing guide
This adds a new integration testing guide to help contributors better
understand how CDK integration testing works and under what scenarios
integration tests should be added. This is not meant to be a
comprehensive guide and can serve as a starting point for us to add
additional scenarios, etc.

This also adds a checklist item to the pull request template for
contributors to indicate that their new feature contains a corresponding
integration test and that they have followed the integration test guide.