chore(v2): only advertise alpha packages with L2s (#19440)
Right now, we advertise `alpha` packages for every service module that
is marked as `experimental`.

However, all L1-only packages are all always marked as `experimental` as
a protection against human mistakes: so that when we add in L2s we don't
accidentally mark them as `stable` right away.

This leads to every CFN-only package having an advertising clause for
an `alpha` package that doesn't even exist.

Do a source code check to only advertise if it looks like there are
actual L2s.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*