fix(apigateway): missing rmethod responses and role in Step Functions integration
The method responses and role were only automatically created when using
the `StepFunctionsRestApi` construct. Move the logic inside the
the integration.

It's now possible to do:

```ts
api.root.addResource('sfn').addMethod('POST', StepFunctionsIntegration.startExecution(stateMachine));
```

Previously this did not create the proper method responses and required
a role to be passed.