chore(cli): integ test filtering is broken (#19462)
Turns out that Jest's `test.concurrent()` does not play well when
running `jest -t TESTNAME`: jest will actually start all tests,
not just the one you were trying to run.

Reported here: https://github.com/facebook/jest/issues/12588

We used to have something to opt-out of concurrent running: we
have to flip that around, and opt in to concurrent running only
for the canary and pipeline tests, where we *know* for a fact
we're not filtering.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*