docs: add integration testing guide (#19469)
This adds a new integration testing guide to help contributors better
understand how CDK integration testing works and under what scenarios
integration tests should be added. This is not meant to be a
comprehensive guide and can serve as a starting point for us to add
additional scenarios, etc.

This also adds a checklist item to the pull request template for
contributors to indicate that their new feature contains a corresponding
integration test and that they have followed the integration test guide.


----

### All Submissions:

* [X] Have you followed the guidelines in our [Contributing guide?](../CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* ~[ ] This PR adds new unconventional dependencies following the process described [here](../CONTRIBUTING.md/#adding-new-unconventional-dependencies)~

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*