fix(codepipeline): cannot deploy pipeline stack with crossAccountKeys twice (under feature flag)
When multiple copies of the same pipeline are deployed in separate stacks, the alias name for the KMS key is the same, causing the deployment to fail. This hcange fixes that using the stack name instead of the stack ID to create a stack safe uniqueId for the alias name. This fix is behind the following feature flag:

@aws-cdk/aws-codepipeline:crossAccountKeyAliasStackSafeUniqueId

Fixes issue #18828.