chore(cdk-cli-wrapper): create a cli wrapper for invoking the cdk
This is initially needed for the integration testing project. For
integration testing we need to be able to programmatically invoke the
CLI and control which arguments are passed to the CLI.

We could have just included this as part of the integration testing
library that we will be creating, but since the logic is pretty self
contained this allows us to eventually publish this as a separate
library for customers to use. Initially we will not publish this
externally, but it is configured with jsii so we can in the future.

This PR adds the base functionality needed, and the goal will be to
iterate on this as we work on creating the integration testing library.

I have taken the command options from the aws-cdk package, which does
introduce some duplication. Ideally it would be cool if we could define
these in a single place and somehow autogenerate the cli args from the
interfaces. Since we don't add new CLI options that often, I think this
is fine for now.

The `ICdk` interface is similar to the [IAws](https://github.com/aws/aws-cdk/blob/fe0ecb48d98f3a33237ae80664610611b90873a9/packages/cdk-assets/lib/aws.ts#L6)
interface in `cdk-assets` which will allow us to create different
"clients" (e.g. we could create another one in `aws-cdk` that just uses
the functions there).