chore(cloudwatch): make constructor private on Color utility class  (#16613)
The `Color` has a public constructor by default, even though it is a utility class that is just used for its static methods. This causes the docs to generate a code snippet for initializing the class, which is misleading to users.

Changing the constructor to private shouldn't be a breaking change since constructing the class didn't do anything in the past.

(docs link: https://docs.aws.amazon.com/cdk/api/latest/docs/@aws-cdk_aws-cloudwatch.Color.html)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*