feat: integ-runner [WIP]
Creating this PR so that we can iterate on the design of the integration
test runner.

We will create two packages (names up for debate)

1. `@aws-cdk/integ-runner` (this package)
CLI tool that will execute integration tests

2. `@aws-cdk/integ-tests`
Library that is used to create integration test cases. This library will
synthesize an assembly file that will be consumed by the runner.