feat: add new integration test runner (#19529)
Creating this PR so that we can iterate on the design of the integration
test runner.

We will create two packages (names up for debate)

1. `@aws-cdk/integ-runner` (this PR)
CLI tool that will execute integration tests

2. `@aws-cdk/integ-tests` (future PR)
Library that is used to create integration test cases. This library will
synthesize an assembly file that will be consumed by the runner.

This PR does not yet implement running new style integration tests (generated from the `@aws-cdk/integ-tests` library).
The initial implementation of the runner will only work on the legacy (i.e. existing) integration tests. 


----

### All Submissions:

* [ ] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/master/INTEGRATION_TESTS.md)?
	* [ ] Did you use `cdk-integ` to deploy the infrastructure and generate the snapshot (i.e. `cdk-integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*