chore: throw more descriptive error when failing to create output directory (#18296)
I passed a `--output` to the CDK that was `/public/something/or/other`, on a system with no `/public`, and then forgot about it. I was presented with the head-scratch-inducing failure message:

```
ENOENT: no such file or directory, mkdir '/public'
Error: ENOENT: no such file or directory, mkdir '/public'
```

This should improve the error so that some reference is made to why the CDK expects or desires this path to exist (and remind me that in fact I told it to go there).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*