fix(eks): incorrect version of aws-node-termination-handler (#19510)
A previous [PR](https://github.com/aws/aws-cdk/pull/18841) upgraded the version of `aws-node-termination-handler`. However, the version contained in the change, 1.14.0, refers to the `appVersion` field and not the `version` field. This change upgrades it to the latest available `version`, 0.18.0, as shown by the command:

```
$ helm search repo aws-node-termination-handler
NAME                            	CHART VERSION	APP VERSION	DESCRIPTION
eks/aws-node-termination-handler	0.18.0       	1.16.0     	A Helm chart for the AWS Node Termination Handler.
```

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md)

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*