chore(aws-lambda): fix broken integration tests
This updates/removes any reference to the deprecated NODEJS_10 runtime
since you can no longer deploy functions with that runtime. It also
updates the docker image used in the docker integ tests to use an image
from the public ecr repository so that it is not tied to a specific
region's ECR repo.