fix(cli): support attributes of DynamoDB Tables for hotswapping (#19620)
When using Nested Stacks, Parameters are used to pass values between the parent and child Stacks.
A customer had a setup where a DynamoDB Table was in the parent Stack,
and a Lambda using that Table (through environment variables) was in the child Stack.
In order to be able to hotswap the Lambda in the child Stack,
we have to correctly resolve attributes of the `AWS::DynamoDB::Table` resource,
like `Ref` and `Arn`, to resolve the values of the Parameters passed to the child Stack.

Fixes #19421

----

### All Submissions:

* [x] Have you followed the guidelines in our [Contributing guide?](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md)

### Adding new Unconventional Dependencies:

* [ ] This PR adds new unconventional dependencies following the process described [here](https://github.com/aws/aws-cdk/blob/master/CONTRIBUTING.md/#adding-new-unconventional-dependencies)

### New Features

* [ ] Have you added the new feature to an [integration test](https://github.com/aws/aws-cdk/blob/master/INTEGRATION_TESTS.md)?
	* [ ] Did you use `cdk-integ` to deploy the infrastructure and generate the snapshot (i.e. `cdk-integ` without `--dry-run`)?

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*