fix(ecs): get rid of EFS casing warnings
About a year ago, ECS TaskDefinition handler changed the casing of some
EFS-related properties:

* `EfsVolumeConfiguration` -> `EFSVolumeConfiguration`
* `FileSystemId` -> `FilesystemId`

They continue to accept both casings, but emit a warning when the
deprecated casing is used. When the new casing was introduced, we
reverted to the old casing in order to not cause resource replacements.

However:

- The old casings emit warnings; when the service/task creation fails
  due to unrelated reasons, users see the warnings, interpret them as
  errors, then stop looking and come and tell us that there is a bug
  in CDK.
- Task definition replacement isn't actually a problem. Task definitions
  can be replaced for something as trivial as changing CPU count or
  memory size. Replacing them for a change that is effectively a no-op
  shouldn't matter. Yes, this will restart `Service`s based on these
  Task Definitions, but if you are only running 1 copy of the Task
  you have made the decision not to care about potential downtime of
  your service.

Maintaining the patch does not seem worth the cost/benefit ratio.

Reverts #10483, closes #15025.